witness generation benchmark
